task2_databases/requirements.txt
psycopg2
PyYAML
